requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
requests==2.31.0
pydantic==1.10.14
shapely==2.0.3
Pillow==10.2.0
